Move styles from lumo to common 2253
