+++ jmix-flowui/flowui-kit/src/main/resources/META-INF/frontend/theme/material/twin-column/jmix-twin-column.js
@@ -14,6 +14,4 @@
  * limitations under the License.
  */
 
-import '@vaadin/input-container/theme/material/vaadin-input-container.js';
-import './jmix-twin-column-styles.js';
 import '../../../src/twin-column/jmix-twin-column.js';
