+++ jmix-flowui/flowui/src/main/resources/io/jmix/flowui/messages.properties
@@ -199,11 +199,10 @@ upload.file.start=Start
 upload.file.remove=Remove
 upload.units=B, kB, MB, GB, TB, PB, EB, ZB, YB
 
-twinColumn.selectItems=Select marked items
-twinColumn.selectAllItems=Select all items
-twinColumn.deselectItems=Deselect marked items
-twinColumn.deselectAllItems=Deselect all items
-twinColumn.clearColumnsSelection=Clear items selection in columns
+twinColumn.selectItems.tooltip=Select marked items
+twinColumn.selectAllItems.tooltip=Select all items
+twinColumn.deselectItems.tooltip=Deselect marked items
+twinColumn.deselectAllItems.tooltip=Deselect all items
 
 io.jmix.flowui.app.multivaluepicker/multiValueSelectDialog.title=Select value
