+++ jmix-translations/content/io/jmix/flowui/messages_ru.properties
@@ -199,11 +199,10 @@ upload.file.start=Возобновить
 upload.file.remove=Удалить
 upload.units=B, kB, MB, GB, TB, PB, EB, ZB, YB
 
-twinColumn.selectItems=Выбрать отмеченные элементы
-twinColumn.selectAllItems=Выбрать все элементы
-twinColumn.deselectItems=Отменить выбор отмеченных элементов
-twinColumn.deselectAllItems=Отменить выбор всех элементов
-twinColumn.clearColumnsSelection=Очистить выбор элементов в колонках
+twinColumn.selectItems.tooltip=Выбрать отмеченные элементы
+twinColumn.selectAllItems.tooltip=Выбрать все элементы
+twinColumn.deselectItems.tooltip=Отменить выбор отмеченных элементов
+twinColumn.deselectAllItems.tooltip=Отменить выбор всех элементов
 
 io.jmix.flowui.app.multivaluepicker/multiValueSelectDialog.title=Выбор значения
